+++ hsapi/dbaccess/db_records.php
@@ -336,7 +336,7 @@ function recordAdd($system, $record, $return_id_only=false){
     }else {
 
         array_shift( $owner_grps ); //remove first
-        if($access_grps!=null || count($owner_grps)>0){
+        if($access_grps!=null || (is_array($owner_grps) && count($owner_grps)>0)){
             updateUsrRecPermissions($mysqli, $newId, $access_grps, $owner_grps);
         }
 
@@ -647,6 +647,7 @@ function recordSave($system, $record, $use_transaction=true, $suppress_parent_ch
 
         $stmt = $mysqli->prepare($query);
         
+        //Call the $stmt->bind_param() method with atrguments (string $types, mixed &...$vars)
         call_user_func_array(array($stmt, 'bind_param'), referenceValues($params));
 
         if(!$stmt->execute()){
@@ -813,7 +814,7 @@ function recordSave($system, $record, $use_transaction=true, $suppress_parent_ch
             //1. cfn_RecTypeIDs -> cfn_ID
             //2. defRecStructure where rst_CalcFunctionID  -> rst_RecTypeID+rst_DetailTypeID 
             $aff_rectypes = findAffectedCalcFields( $system, $rectype );
-            if($aff_rectypes!=null && count($aff_rectypes)>0){
+            if(is_array($aff_rectypes) && count($aff_rectypes)>0){
                 recordUpdateCalcFields( $system, null, $aff_rectypes);    
             }
             
@@ -828,15 +829,15 @@ function recordSave($system, $record, $use_transaction=true, $suppress_parent_ch
         $mask = mysql__select_value($mysqli,"select rty_TitleMask from defRecTypes where rty_ID=".RT_RELATION);
 
         $relRecs = recordGetRelationship($system, $recID, null, array('detail'=>'ids'));
-        if($relRecs!=null && count($relRecs)>0){
+        if(is_array($relRecs) && count($relRecs)>0){
             $relRecsIDs = $relRecs;
         }
         $relRecs = recordGetRelationship($system, null, $recID, array('detail'=>'ids'));
-        if($relRecs!=null && count($relRecs)>0){
+        if(is_array($relRecs) && count($relRecs)>0){
             $relRecsIDs = array_merge($relRecsIDs, $relRecs);
         }
         //reset temporary flag for all relationship records
-        if(count($relRecsIDs)>0){
+        if(is_array($relRecsIDs) && count($relRecsIDs)>0){
             foreach($relRecsIDs as $relID){
                 $res = recordUpdateTitle($system, $relID, $mask, 'Title Mask for Relationship not defined');
             }
@@ -846,7 +847,7 @@ function recordSave($system, $record, $use_transaction=true, $suppress_parent_ch
 
         //recordGetLinkedRecords - get all linked and related records and update them
         $links = recordGetLinkedRecords($system, $recID);
-        if($links!=null && count($links)>0){
+        if(is_array($links) && count($links)>0){
             //find title masks
             $links_rectypes = array_unique(array_values($links));
             $masks = mysql__select_assoc2($mysqli,'select rty_ID, rty_TitleMask from defRecTypes where rty_ID in ('
@@ -970,7 +971,8 @@ function recordDelete($system, $recids, $need_transaction=true,
         if($check_source_links && count($allowed_recids)>0){
             $links = recordSearchRelated($system, $allowed_recids, -1, 'ids', 1);
 
-            if($links['status']==HEURIST_OK && @$links['data']['reverse']!=null && count(@$links['data']['reverse'])>0){
+            if($links['status']==HEURIST_OK && @$links['data']['reverse']!=null 
+                && is_array(@$links['data']['reverse']) && count($links['data']['reverse'])>0){
                 return array('status'=>HEURIST_OK, 
                     'data'=> array( 'source_links_count'=>count($links['data']['reverse']),
                         'source_links'=>implode(',',$links['data']['reverse']) ));
@@ -1144,7 +1146,7 @@ function recordUpdateOwnerAccess($system, $params){
         $owner_grps = prepareIds( @$params['OwnerUGrpID'], true);
         $access = @$params['NonOwnerVisibility'];
 
-        if($owner_grps==null || count($owner_grps)==0 || $access==null){             
+        if(!is_array($owner_grps) || count($owner_grps)==0 || $access==null){             
             return $system->addError(HEURIST_INVALID_REQUEST, 'Neither owner nor visibility parameters defined');
         }
 
@@ -1157,7 +1159,7 @@ function recordUpdateOwnerAccess($system, $params){
             $recids = mysql__select_list2($mysqli, 'SELECT rec_ID from Records where rec_ID in ('
                 .implode(',', $recids).') and rec_RecTypeID='. $rec_RecTypeID);
 
-            if($recids==null || count($recids)==0){             
+            if(!is_array($recids) || count($recids)==0){             
                 return $system->addError(HEURIST_NOT_FOUND, 'No record found for provided record type');
             }
         }
@@ -1555,7 +1557,7 @@ function addParentToChildPointer($mysqli, $child_id, $child_rectype, $parent_id,
             .'AND rst_DetailTypeID=dty_ID';
 
             $pointers = mysql__select_assoc2($mysqli, $query);
-            if(count($pointers)>0)
+            if(is_array($pointers) && count($pointers)>0)
                 foreach($pointers as $dt_ID=>$ptr){
                     if($ptr) $ptr = explode(',',$ptr);
                     if(count($ptr)>0 && in_array($child_rectype, $ptr)){
@@ -1656,7 +1658,7 @@ function recordCanChangeOwnerwhipAndAccess($system, $recID, &$owner_grps, &$acce
 
     //$ownerid_old = @$record["rec_OwnerUGrpID"]; //current ownership
     //new owners are not defined - take current one
-    if($owner_grps==null || count($owner_grps)==0 || !($owner_grps[0]>=0)){  
+    if(!is_array($owner_grps) || count($owner_grps)==0 || !($owner_grps[0]>=0)){  
         $owner_grps = $current_owner_groups;
     }
     if(array_search(0, $owner_grps, true)!==false){ //there is "everyone" 
@@ -1748,15 +1750,15 @@ function findAffectedCalcFields( $system, $rty_ID ){
     
     $mysqli = $system->get_mysqli();
 
-    $query = 'SELECT cfn_ID FROM defCalcFunctions WHERE find_in_set('.$rty_ID.',cfn_RecTypeIDs) <> 0';
+    $query = 'SELECT cfn_ID FROM defCalcFunctions WHERE find_in_set('.$mysqli->real_escape_string($rty_ID).',cfn_RecTypeIDs2) <> 0';
     $field_ids = mysql__select_list2($mysqli, $query);
 
-    if(count($field_ids)>0){
+    if(is_array($field_ids) && count($field_ids)>0){
         
         $query = 'SELECT rst_RecTypeID WHERE rst_CalcFunctionID IN ('.implode(',',$field_ids).')';
         $rectype_ids = mysql__select_list2($mysqli, $query);
         
-        if(count($rectype_ids)>0){
+        if(is_array($rectype_ids) && count($rectype_ids)>0){
             recordUpdateCalcFields($system, null, $rectype_ids);
         }
     }
@@ -1785,8 +1787,8 @@ function recordUpdateCalcFields($system, $recID, $rty_ID=null, $progress_session
         if ($res){
             while ($row = $res->fetch_row()){
                 if($rty_ID != $row[0]){
-                    if($rty_ID){
-                        $rec_count = $rec_count + count(@$rectypes[$rty_ID]);   
+                    if($rty_ID && is_array(@$rectypes[$rty_ID])){
+                        $rec_count = $rec_count + count($rectypes[$rty_ID]);   
                     }
                     $rty_ID = $row[0];
                     $rectypes[$rty_ID] = array();
@@ -1795,7 +1797,9 @@ function recordUpdateCalcFields($system, $recID, $rty_ID=null, $progress_session
             }
             $res->close();
         }
-        if($rty_ID) $rec_count = $rec_count + count(@$rectypes[$rty_ID]);
+        if($rty_ID && is_array(@$rectypes[$rty_ID])){
+          $rec_count = $rec_count + count($rectypes[$rty_ID]);  
+        } 
         
     }else if($recID>0){
         
@@ -1817,7 +1821,7 @@ function recordUpdateCalcFields($system, $recID, $rty_ID=null, $progress_session
             $rty_ID = prepareIds($rty_ID);
         }
         
-        if($rty_ID==null || count($rty_ID)==0){
+        if(!is_array($rty_ID) || count($rty_ID)==0){
             //all rectypes - entire database
             $rty_ID = mysql__select_list2($mysqli, 'SELECT rty_ID FROM defRecTypes'); 
             $rec_count = mysql__select_value($mysqli, 'SELECT count(rec_ID) FROM Records WHERE (NOT rec_FlagTemporary)'); 
@@ -1859,13 +1863,13 @@ function recordUpdateCalcFields($system, $recID, $rty_ID=null, $progress_session
             .' AND cfn_ID=rst_CalcFunctionID');
             
         //there are not calculation fields for this record type
-        if(!$formulae || count($formulae)==0){ 
+        if(!is_array($formulae) || count($formulae)==0){ 
             
             if($record_ids=='*'){
                $cnt = mysql__select_value($mysqli, 'SELECT count(rec_ID) FROM Records '
                 .'WHERE (rec_RecTypeID='.$rty_ID.') AND (NOT rec_FlagTemporary)'); 
                $progress_count = $progress_count + $cnt;
-            }else{
+            }else if (is_array($record_ids)) {
                $progress_count = $progress_count + count($record_ids);
             }
             
@@ -1903,7 +1907,7 @@ function recordUpdateCalcFields($system, $recID, $rty_ID=null, $progress_session
                          break;
                      }
                 }else{
-                    if($idx<count($record_ids)){
+                    if(is_array($record_ids) && $idx<count($record_ids)){
                          $recID = $record_ids[$idx];
                          $idx++;
                     }else{
@@ -2047,7 +2051,7 @@ function executeSmarty($system, $params, $mode=null, $heuristRec=null){
   
   $record_ids = @$params['records'];
   
-  if($record_ids==null || count($record_ids)<1) return '';
+  if(!is_array($record_ids) || count($record_ids)<1) return '';
   
   $mode = $mode ?$mode:'eval:'; //string: - use complied or eval: - compile every time
 
@@ -2771,14 +2775,14 @@ function _prepareDetails($system, $rectype, $record, $validation_mode, $recID, $
 
         $system->addError(HEURIST_ACTION_BLOCKED, $sMsg, null);
 
-    }else if (count($det_required)>0) {
+    }else if (is_array($det_required) && count($det_required)>0) {
 
         $system->addError(HEURIST_ACTION_BLOCKED, 'Required field'.(count($det_required)>1?'s':'')
             .' missing value or '.
             (count($det_required)>1?'have':'has')
             .' invalid value: '.implode(',',array_values($det_required)));
 
-    }else if (count($insertValues)<1) {
+    }else if (!is_array($insertValues) || count($insertValues)<1) {
 
         $system->addError(HEURIST_INVALID_REQUEST, "Fields are not defined");
 
@@ -3052,20 +3056,23 @@ function updateUsrRecPermissions($mysqli, $recIDs, $access_grps, $owner_grps){
 
     $recIDs = prepareIds($recIDs);
 
-    if(count($recIDs)>0){
+    if(is_array($recIDs) && count($recIDs)>0){
 
         $query = 'DELETE FROM usrRecPermissions WHERE rcp_RecID in ('.implode(',', $recIDs).')';
         $mysqli->query($query);
 
         $access_grps = prepareIds($access_grps);
         $owner_grps = prepareIds($owner_grps, true);
-        if(count($access_grps)>0 || count($owner_grps)>0){
+        if( (is_array($access_grps) && count($access_grps)>0) 
+            || (is_array($owner_grps) && count($owner_grps)>0)){
             //add group record permissions
             $values = array();
             foreach($recIDs as $recID){
+                if(is_array($owner_grps))
                 foreach ($owner_grps as $grp_id){
                     array_push($values,'('.$grp_id.','.$recID.',"edit")');
                 }
+                if(is_array($access_grps))
                 foreach ($access_grps as $grp_id){
                     array_push($values,'('.$grp_id.','.$recID.',"view")');
                 }
